pyproject.toml
[project]
name = "prodigy-prot"
license = "Apache-2.0"
version = "2.4.0"
description = "PROtein binDIng enerGY prediction"
authors = [
  { name = "Anna Vangone" },
  { name = "Joao Rodrigues" },
  { name = "Joerg Schaarschmidt" },
]
maintainers = [{ name = "BonvinLab", email = "bonvinlab.support@uu.nl" }]
readme = "README.md"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = ["biopython>=1.80", "freesasa>=2.2.1", "numpy>=2"]

[project.optional-dependencies]
dev = ["pytest", "coverage", "hypothesis", "pytest-cov", "mypy"]

[project.scripts]
prodigy = "prodigy_prot.cli:main"

[tool.setuptools]
include-package-data = true
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = ["integration: marks tests as integration tests"]

[tool.mypy]
disable_error_code = ["import-not-found"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
